requirements.txt
Flask
pymongo==3.11
gunicorn
tensorflow
pillow
numpy
deeplake<4
gdown